Add LocationManager for the free version Remove PhotoView dependency and from About's string
