+++ app/build.gradle
@@ -31,7 +31,5 @@ android {
 dependencies {
     compile 'com.android.support:appcompat-v7:21.0.3'
     compile 'com.android.support:cardview-v7:21.0.3'
-    // Version 1.2.3 has a bug and crashes when detecting 2-finger gestures. Use 1.2.2 instead.
-    compile 'com.github.chrisbanes.photoview:library:1.2.2'
     unfreeCompile 'com.google.android.gms:play-services-location:6.5.87'
 }
