+++ app/src/main/res/drawable/accuracy_background.xml
