+++ app/src/main/res/values/strings.xml
@@ -115,6 +115,7 @@
 \nThe name FOSDEM and the gear logo are registered trademarks of FOSDEM VZW.  Used with permission.
 \n
 \n<small>This application makes use of the following libraries:
-\n- <a href="http://developer.android.com/tools/support-library/">Android Support Library</a><i> by The Android Open Source Project</i>\n- <a href="http://viewpagerindicator.com/">ViewPagerIndicator</a><i> by Jake Wharton</i>\n- <a href="https://github.com/chrisbanes/PhotoView">PhotoView</a><i> by Chris Banes</i></small></string>
+\n- <a href="http://developer.android.com/tools/support-library/">Android Support Library</a><i> by The Android Open Source Project</i>\n- <a href="http://viewpagerindicator.com/">ViewPagerIndicator</a><i> by Jake Wharton</i></small></string>
+    <string name="your_position">Your position</string>
 
 </resources>
