--- app/src/unfree/AndroidManifest.xml
@@ -1,5 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
-
-</manifest>
