--- app/src/unfree/res/values/colors.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="precision_color">#88a7168e</color>
-    <color name="precision_stroke_color">#a7168e</color>
-</resources>
