--- app/src/unfree/res/values/dimens.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <dimen name="position_dot_size">5dp</dimen>
-</resources>
