--- app/src/unfree/res/values/strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="your_position">Your position</string>
-</resources>
